+++ opps/bin/opps-admin.py
@@ -1,19 +1,21 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-import sys
+import argparse
 from django.conf import settings
 from django.core import management
 
 
 settings.configure()
 if __name__ == "__main__":
-    if sys.argv[1] == 'startproject':
-        if len(sys.argv) < 2:
-            raise management.CommandError(
-                "accepet accepts one argument - the name of the project "
-                "to create.")
+    parser = argparse.ArgumentParser(description='Opps CMS bin file')
+    parser.add_argument('operation', help='task to be performed',
+                        choices=['startproject'])
+    parser.add_argument("project_name", help="Project name", type=str)
+
+    args = parser.parse_args()
+    if args.operation == 'startproject':
         management.call_command(
-            'startproject', sys.argv[2],
+            'startproject', args.project_name,
             template='https://github.com/opps/opps-project-template/zipball/'
             'master',
             extensions=('py', 'md', 'dev')